Fixed nextauth typeerror and removed some unused imports. Build works now
