+++ next/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@auth/prisma-adapter": "^1.0.6",
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^5.4.2",
     "@types/react": "18.2.21",
     "@types/react-dom": "18.2.7",
